Refactor: rename getCalendardata and extract useGetCalendarData
