+++ src/components/coffee/HistoryTab.tsx
@@ -1,11 +1,7 @@
-import { useQuery } from '@tanstack/react-query';
-import { useRecoilValue } from 'recoil';
-
 import MyCalendar from '@/components/coffee/MyCalendar';
 import CaffeineHistory from '@/components/coffee/CaffeineHistory';
 
-import { getCoffeeCaledar } from '@/api/coffee';
-import { activeMonthState } from '@/atoms/atoms';
+import { useGetCalendarData } from '@/hooks/coffee/useGetCalendarData';
 
 const HistoryTab = ({
   signedIn,
@@ -14,15 +10,7 @@ const HistoryTab = ({
   signedIn: string;
   activeStartDate: Date;
 }) => {
-  const activeMonth = useRecoilValue(activeMonthState);
-  const { data } = useQuery({
-    queryKey: ['coffeeCalendar', activeMonth],
-    queryFn: async () => {
-      const data = await getCoffeeCaledar(activeMonth);
-      return data;
-    },
-    enabled: !!signedIn
-  });
+  const { data, activeMonth } = useGetCalendarData(signedIn);
 
   return (
     <>
